requirement.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
black==23.12.0
click==8.1.7
Django==5.0
django-admin-interface==0.27.0
django-colorfield==0.11.0
django-decouple==2.1
django-tinymce==3.6.1
frozenlist==1.4.1
idna==3.6
isort==5.13.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
python-decouple==3.8
python-slugify==8.0.1
ruff==0.1.8
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.9.0
yarl==1.9.4
